requirements.txt
google-generativeai
pydantic
fastapi
uvicorn[standard]
python-dotenv
python-multipart
aiofiles # Often needed by FastAPI for async file operations
jinja2 # For HTML templating with FastAPI
